Document directive regex
